Fix coordinates system
Mediapipe returns landmark with values from 0 to 1, with x based on the frame's width and y on the frame's height. We used whose values as if they were based on the same value so the computation were wrong. Now everything is based on pixels of the frame.
All configurations now need to be adjusted because the x/y ratio are now different.